refactor: introduce BoundedCache base class and refactor BoundedExpressionCache and BoundedFilterCache to inherit from it
