--- test/Kista.Testing.TUnit/Fixtures/IRelationship.cs
@@ -1,7 +1,0 @@
-﻿namespace Kista {
-	public interface IRelationship {
-		string Type { get; set; }
-
-		string FullName { get; set; }
-	}
-}
